rules_modifier
We would like you to write a script that matches rules from a security
group for EC2 and update it. For example I have a security group called
"database-servers" that have many rules. We want all rules that has the
source "100.100.100.100/32" to be updated to "200.200.200.200/32". You
may code it in a your language of choice. Assume the user will be
running it on the command line.